# src/main/resources/static/swagger-ui/openapi.yaml
openapi: 3.0.0
info:
  title: API de Residencias
  version: 1.0.0
  description: API para gestionar residencias, con operaciones de registro y consulta.
paths:
...
  /resi/{idResidencia}/juego/{idJuego}/registro/add:
    post:
      tags:
        - Modulo - JUEGO
      summary: Registrar un nuevo registro de juego.
      description: Crea un nuevo registro de juego validando campos obligatorios,existencia de las referencias y valores validos.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroJuego.Dto'
      responses:
        '201':
          description: Registro de juego registrado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistroJuego.ResponseDto'
        '400':
          description: Datos nulos/vacios .
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00

        '404':
          description: Juego, Residente o Usuario invalido, no esixte ninguna referencia en la base de datos de estas entidades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                RESIDENTE_INVALIDO:
                  value:
                    mensaje: Residente invalida
                    codigo: 1009
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                USUARIO_INVALIDO:
                  value:
                    mensaje: Usuario invalida
                    codigo: 1010
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
        '406':
          description: No valores negativos en el el campo de puntos/num
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                VALORES_NEGATIVOS:
                  value:
                    mensaje: No puede ser valores negativos
                    codigo: 1007
                    status: 406
                    timestamp: 2025-04-13T21:00:00

        '409':
          description: Conflicto con las referencias de la residencia, no coinciden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CONFLICTO_REFERENCIAS:
                  value:
                    mensaje: Problemas con las referencias de las entidades, no corresponden a las mismas
                    codigo: 1011
                    status: 409
                    timestamp: 2025-04-13T21:00:00
  /resi/{idResidencia}/juego/{idJuego}/registro/{idRegistroJuego}/get:
    get:
      tags:
        - Modulo - JUEGO
      summary: Obtener registro de juego
      description: Devuelve los datos de una partida específica.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
        - name: idRegistroJuego
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Registro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistroJuego.ResponseDto'
        '400':
          description: Datos nulos/vacios .
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: juego/ registro_juego invalido, no existe en la base de datos.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                REGISTRO_JUEGO_INVALIDO:
                  value:
                    mensaje: Registro Juego invalido
                    codigo: 1012
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:00
...
  /resi/{idResidencia}/juego/{idJuego}/registro/{idRegistroJuego}/addComment:
    patch:
      tags:
        - Modulo - JUEGO
      summary: Añadir comentario a registro
      description: Actualiza la observación/comentario del registro.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
        - name: idRegistroJuego
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroJuego.UpdateDto'
      responses:
        '200':
          description: Comentario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistroJuego.ResponseDto'
        '400':
          description: Datos nulos/vacios .
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: juego/ registro_juego invalido, no existe en la base de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                REGISTRO_JUEGO_INVALIDO:
                  value:
                    mensaje: Registro Juego invalido
                    codigo: 1012
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:00
  /resi/{idResidencia}/juego/{idJuego}/registro/{idRegistroJuego}/delete:
    delete:
      tags:
        - Modulo - JUEGO
      summary: Eliminar registro de juego
      description: Elimina un registro de juego de forma permanente.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
        - name: idRegistroJuego
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Registro eliminado correctamente
        '400':
          description: Datos nulos/vacios .
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: juego/ registro_juego invalido, no existe en la base de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                REGISTRO_JUEGO_INVALIDO:
                  value:
                    mensaje: Registro Juego invalido
                    codigo: 1012
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:00
...
components:
  schemas:
...
    RegistroJuego.Dto:
      type: object
      properties:
        idResidente:
          type: integer
          format: int64
          example: 5
        idUsuario:
          type: integer
          format: int64
          example: 3
        num:
          type: integer
          example: 2
        duracion:
          type: number
          format: double
          example: 120.5
        dificultad:
          type: string
          enum: [ FACIL, MEDIA, DIFICIL ]
          example: MEDIA
    RegistroJuego.UpdateDto:
      type: object
      properties:
        observacion:
          type: string
          example: Muy lento
    RegistroJuego.ResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 42
        idResidente:
          type: integer
          format: int64
          example: 5
        idJuego:
          type: integer
          format: int64
          example: 1
        idUsario:
          type: integer
          format: int64
          example: 3
        num:
          type: integer
          example: 2
        duracion:
          type: number
          format: double
          example: 120.5
        dificultad:
          type: integer
          format: int64
          example: 1
        fecha:
          type: string
          format: date-time
          example: "2025-04-16T10:15:30"
...
    Error.ResponseDto:
      type: object
      properties:
        mensaje:
          type: string
        codigo:
          type: integer
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        mensaje: Faltan campos obligatorios
        codigo: 1001
        status: 400
        timestamp: 2025-04-13T21:00:00
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT